+++ workspaces/adoption-insights/plugins/adoption-insights-backend/src/domain/EventBatchProcessor.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { LoggerService } from '@backstage/backend-plugin-api/index';
+import { LoggerService } from '@backstage/backend-plugin-api';
 import { EventDatabase } from '../database/event-database';
 import { Event } from '../models/Event';