+++ workspaces/lightspeed/plugins/lightspeed-backend/config.d.ts
@@ -38,7 +38,7 @@ export interface Config {
       token?: string;
     }>;
     /**
-     * If turn on query restriction validation.
+     * query restriction validation (Enabled by default).
      * @visibility frontend
      */
     questionValidation?: boolean;