+++ workspaces/adoption-insights/plugins/adoption-insights/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @red-hat-developer-hub/backstage-plugin-adoption-insights
 
+## 0.1.0
+
+### Minor Changes
+
+- a66e0b0: Backstage version bump to v1.36.1
+
+### Patch Changes
+
+- Updated dependencies [a66e0b0]
+  - @red-hat-developer-hub/backstage-plugin-adoption-insights-common@0.2.0
+
+## 0.0.4
+
+### Patch Changes
+
+- 16ecc8f: Updated dependency `@mui/lab` to `5.0.0-alpha.176`.
+
 ## 0.0.3
 
 ### Patch Changes